Fixed gazebo robot spawn problem
The problem was that first in the launch file, the robot spawner was called wrong. That is fixed but the robot didn't appeared in Gazebo. This lead to the second problem which is that the inertial part in the URDF file was missing. (URDF consist of two main parts: visual and inertial. With only visual part, you can visualize only in Rviz. To spawn the robot in Gazebo, you need the inertial part also). I took it from the mybot_description. You can make your robot model changes on forkbot.xacro now.